Refactor Celery tasks and admin for scheduled video publishing
