--- pod/__init__.py
@@ -1,5 +1,0 @@
-from __future__ import absolute_import, unicode_literals
-
-from .celery import app as celery_app
-
-__all__ = ('celery_app',)
